+++ CHANGELOG.md
@@ -9,6 +9,7 @@ HEAD
 * Change the default of `ARDUINOJSON_USE_DOUBLE` to `1`
 * Change the default of `ARDUINOJSON_USE_LONG_LONG` to `1` on 32-bit platforms
 * Add `as<JsonString>()` and `is<JsonString>()`
+* Add safe bool idiom in `JsonString`
 
 v6.18.5 (2021-09-28)
 -------